premisys/requirements.txt
aiohttp==3.6.2
amqp==2.5.2
asgiref==3.2.5
async-timeout==3.0.1
attrs==19.3.0
billiard==3.6.3.0
celery==4.4.2
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.4
dnspython==1.16.0
idna==2.8
kombu==4.6.8
multidict==4.7.3
pytz==2019.3
sqlparse==0.3.1
vine==1.3.0
yarl==1.4.2
djangorestframework==3.11.0
pyodbc==4.0.30
#pypyodbc==1.3.5
psycopg2-binary==2.8.5
gunicorn==20.0.4
django-cors-headers==3.4.0
XlsxWriter==1.0.2
meraki-sdk==1.5.0


